Don't overwrite createvrf flag for mirror config updates
Agent sets create vrf flag to true if it creates VRF as part of mirror
entry addition. when mirror entry is deleted , Agent checks vreate vrf flag,
if it is set then deletes teh VRF. In this scenario, mirror config updates
are overwriting the flag to false and that blocks the VRF deletion

made changes not to overwrite the flag if it set.

Change-Id: I6a202da04a91b1d2d9554c13579e1d960cf92e0e
Closes-bug: #1803264