Change in parsing of VPN name from routing instance
The routing instances don't have any generalized
nomenclature as of now.To support that changing
the parsing method to identify the prefixes of VPN and return the split.
Added filter to stop sending of APPTRACK_SESSION_CLOSE syslogs for incoming traffic.
Fix is identified by checking for "LAN-" prefix in routing_instance field.
Change of Network key to tenant::VPN so that VPN is binded to tenant.

Change-Id: I79881b57aeccf023c73a097185218a83b872ffd0
Closes-Bug: #1810887