Disable log level checks for flows
The flow messages are sent at LOG_INFO by default.
If the configured log level is anything more flow
messages will get dropped. Default configured log
level is SYS_NOTICE. For flow messages to be written
to the syslog, the configured log level has to be
SYS_INFO; but this can cause a huge churn. By
disabling this check only for flow messages, they
will be logged as long as the destination is set as
Syslog.

Change-Id: Ice7b67feeb413620bcc765056835a4221ae91302
Partial-bug: #1811001
Partial-jira-bug: #CBS-7