Chaning the index of route target if going beyond boudary
With 4 byte ASN, index field in route target can only be 2 bytes.
However, schema generates default rtargets with 4 byte value. Real
fix would have been to change those values in schema itself but it
has its own challenges. Here, adding a hack in control node to
change index of a rtarget if asn is 4 byte and index is also 4 byte
(only if in the range of auto generated rtargets)

Change-Id: I40e2fc0fb06bf03bd3c06cd53878dea4f045923f
Partial-Bug: #1807123