Change the location and name of generated certs files
Our cert files have a permission problem in all-in-one setup. During the
provisioning, the cert files with the same name will only be created once,
but they are only accessible for the user who created them. In this case,
neutron is not able to use cert files created by contrail.

To fix the problem, we create one cert file for each user, using the username
as the file's prefix. For discovery cert file, we put it to '/var/lib/contrail/',
which only accessible for contrail user. Also ignore the 'key' and 'cert' file
for discovery, refer to https://review.opencontrail.org/#/c/48064/

Closes Jira Bug: JCB-218812

Change-Id: I5360b6116fc58e84c238bd1b6dd32e21068d44f8