[Config] Temp fix for lost update problem in Security Group Rules
When updating a security group in parallel to add/delete multiple rules,
we see that there's a lost update problem leading to undefined behavior.

This change is to temporarily fix this problem by implementing a
Zookeeper lock inside neutron_plugin_db of API Server.

Long term fix is still in discussion for this.

Change-Id: I1f469db2a550da9d2ff86697848d71b07c00503a
Closes-Jira-Bug: JCB-218976