[Config] Temp fix for lost update problem in Security Group Rules
When updating a security group in parallel to add/delete multiple rules,
we see that there's a lost update problem leading to undefined behavior.

This change is to temporarily fix this problem by implementing a
Zookeeper lock inside neutron_plugin_db of API Server.

Long term fix is still in discussion for this.

Change-Id: I871168e140767e711f1c8ca08a50de321268d304
Closes-Jira-Bug: JCB-218976