[schema transformer] Clean RT when a VN is deleted
API server is now in charge of the creation and deletion of the primary
RI of a VN but the schema stills in charge to allocate a unique route
target number and create/delete the corresponding RT. But actually, when
a VN is deleted, schema tries to delete all referenced RI (in memory and on the
VNC API for all the RI and only in memory for the primary RI). But frequently,
it deletes the primary RI before the VNC API have deleted it and it
tries to delete the allocated RT but that failed as it still referenced
by the primary RI not yet deleted. Then when the primary RI is deleted,
the schema ignore the notification as that RI is not anymore in its memory and
the RT remains (lock in zk was removed and could be re-used).
That patch changes the schema to not delete the primary RI when a VN is
deleted and lets the notification of the primary RI deletion warns the
schema to delete the RI and its allocated RT.

Change-Id: I02ac5c334f9a8029874645b31696fd7af4585907
Closes-Jira-Bug: JCB-219068
(cherry picked from commit 168a92fdad0ba4d8d9866a9671b000053067bd4a)