[vnc openstack] Fix port binding status
VNC OpenStack mapping code sets the port binding vif_type to 'vrouter'
event if the port was not yet allocated to a compute node. Set the port
unbound until it is schedule on compute node.
Fix some issue in the FWaaS mapping code to list VNC resource with
filter.
Fix default FWaaS firewall rule action which was not set and rule was
not enabled.

Change-Id: I3c7865172453339b5c6f44f53d483ec937ee44bc
Partial-Jira-Bug: JCB-204217