[ST]: Fix Global ASN update
Modified global asn update function to re-read the vnc RT object
before updating the RI and logical router refs to it.

CEM-4458 also needs additional fix.
If vxlan routing is enabled, the RTs created for LRs will not have
a back_ref to the routing instance of all the connected VNs.
In ST, currently, we don't have this check and hence when global
ASN is updated, ST adds reference from these RIs to the RT created for LR.

Closes-Jira-Bug: JCB-218989
Closes-Jira-Bug: CEM-4458

Change-Id: I7785d4c028f9ee10bc526744488ea5a6861f2a91