Reworking provision_control.py to avoid overwrite
When this script was called by /provision.sh inside all nodemgr
containers, it was causing overwrite of configuration that was done on
the provisioning stage. Adding check if router was already configured
and skipping provisioning in case it was already done.

This script was very unstructured. This change also does some
refactoring to make it more readable.

Change-Id: I3d6b4dc4e0e5ac3ecc1681139ba224120ef8a4a6
partial-jira-bug: JCB-189626