Add asserts to ensure that UVE key is set before they are sent
Recently a bug was found in uve send where in uve's key was not set during
deletion. In this change, asserts have been added to catch the same, all
other similar places in bgp and xmpp code.

JCB-218965 issue did hit the assert introduced here if the fix were not to be
in place.

Change-Id: If0193726f8d7eb5115135642b3c8b8fd6eb0ae87
Closes-Jira-Bug: JCB-218965