Flow eviction case resulted in memleak in agent UVE Stats module
Agent UVE module adds a new entry to a tree based on UUID during flow creation
and deletes the entry during flow deletion.
Flow eviction results in the flow entry getting re-used with different UUID.
Flow deletion event is not posted. Hence Uve entry is not deletd from the tree.
As part of this fix, when the flow entry gets re-used, before the UUID change
post a delete event to UVE module which deletes the entry from the tree.

Executed the long duration test procedure with and without the fix and verified
the tree is empty once all flows are deleted.

Closes Jira Bug: JCB-218789

Change-Id: I1f8f53d8bbb2bfab33739e9c85294b474e2d34df