[Config] Add support to have different roles to an Logical Router
Before this change, we only allow all the logical routers in a project
to be either work with SNAT or for VXLAN.
This commit is to change that and have a new field in every LR.
A new field called "logical_router_type" will let the user choose
particular LR for either SNAT or VXLAN and with this change the project
level "vxlan_routing" flag will be rendered insignificant.

For upgrade support, code is added that will set the "logical_router_type"
for all the logical routers as per the project level "vxlan_routing" flag.

For backward compatibility, project level "vxlan_routing" flag is left
as it is.
However, in the release containing this fix, "vxlan_routing" flag will
always be set to "False" and changes will be done in UI to flag an error
if user intends to change this flag

Schema changes are done here: https://review.opencontrail.org/#/c/49299/

Change-Id: I47e4e195d642a90fb86f4c879dd17d8815623202
Closes-Jira-Bug: JCB-192482
Depends-On: I5cc96707bf1860e2865ea781c30ab252ac8dace2