[Config] Add support to have different roles to an Logical Router
Before this change, we only allow all the logical routers in a project
to be either work with SNAT or for VXLAN.
This commit is to change that and have a new field in every LR.
A new field called "logical_router_type" will let the user choose particular
LR for either SNAT or VXLAN and with this change the project level
"vxlan_routing" flag will be rendered insignificant.

1. For upgrade support, code is added that will set the "logical_router_type"
for all the logical routers as per the project level "vxlan_routing" flag.

2. For backward compatibility, project level "vxlan_routing" flag is left as it is.
However, in the release containing this fix, "vxlan_routing" flag will always
be set to "False" and changes will be done in UI to flag an error if user
intends to change this flag

3. We will not allow change of "logical_router_type" for an existing LR.
The idea behind that is that a LR created for the purpose of SNAT routing
should be repurposed for the use with VXLAN. User has to instead create
a different LR for that purpose.

Schema changes are done as part of https://review.opencontrail.org/#/c/49299/

Change-Id: I47e4e195d642a90fb86f4c879dd17d8815623202
Closes-Jira-Bug: JCB-192482
Depends-On: I5cc96707bf1860e2865ea781c30ab252ac8dace2