[DM] Fix commit fabric config role
The commit fabric role has a task to retrieve the contents of a file after all the configurations have been loaded. This task fails if the file is not already present (with_file first gets evaluated first and then the 'when' condition is checked)Fixing this issue in this review.
closes-jira-bug: CEM-2773

Change-Id: I836ea12b1c45c8722961d73c288fa0d844ab8001