Config-API: Honor the fields requested during object_read/GET calls
Currently we do an OR of the requested fields and the propery+refs fields
rather that has to be an AND along with back_refs and children fields.
id_perms and perms2 are required for further rbac validation of the object
so fetching those fields irrespective of the requested field.

Change-Id: I7e61be057cb51ae72eaaee7af90ffe95073978bb
Closes-Jira-Bug: JCB-219237