Sending smet capability information to agent
Agent/Vrouter need to know if a peer is smet capable or not.
Route table lookup will depend on that. If a peer is smet
capable then corresponding S, G lookup should happen, otherwise
IMET route lookup should happen. This change is from control
side to propagate this information to agent. Changes are needed
in agent and vrouter to use this

Partial Jira Bug: JCB-219141

Change-Id: Ibdca6c0cd1f564206d4bd97edc95fffd963e5d08