Alter bgpaas port reallocation in ST
There is a change in Cassandra backend for Schema Transformet that
allows for a change of 'start' and 'end' ports in bgpaas port allocator.
This is not implemented in VNC API (reallocate_int_pool method is not
yet implemented). So for now we silently skip this reallocation if ST is
running with etcd/api backend instead of cassandra/rabit pair.

Change-Id: Ic8d8a25c359c03a78ebbe412bf5db22d24e80bb5