Use Boost dynamic libraries
While building unit tests, Boost is compiled into every one of them.
To reduce the size of binaries and the build time we switched to DLL.

For this code to work we have to add BOOST_ALL_DYN_LINK to compilation
flags which will be done in contrail-build.

Partial Jira Bug: WIND-197

Change-Id: Iefc73f4226c8799becc5d7c091075e4feddb69ce