[Config] Handle KeyErrors in uuid_to_fq_name
The SYS_ERR log and KeyError were added as part of
https://bugs.launchpad.net/juniperopenstack/+bug/1771879

However, there is a change to modify the log level as it leads to excessive logs.
As part of this review, instead of raising SYS_ERR for all the KeyErrors,
we will instead raise NoIdError in case fq_name or type is not present in
uuid_to_fq_name function.

This fix is based on comments made by Sachin in https://review.opencontrail.org/#/c/43148/

Closes-Jira-Bug: JCB-206854
Closes-Jira-Bug: JCB-206859
Change-Id: Ic864c4f2d91f6ab804e2ffe9517ce05667cf711b