Add support for multiple static routes for same destination
Currently, static routes are maintained in a map keyed by destination address
prefix. This limits one to configure only one route per destination.

In order to support ECMP, change the key to a combination of prefix an nexthop
so that multiple (ecmp) static routes can be configured for same destination.

Closes-Jira-Bug: JCB-219353
Change-Id: I4949275978b92ea6e97132be49381ca8e0f55208