Fix for pkt drop triggered by restart on non-related empty compute.
Config updates received after the empty compute restart, triggered
revaluation of ARP routes, and consequently the ARP nhs are processed
again. At this point we would see drop on all tunnel nh's which are
marked invalid because of the arp nh processing (arp nh are also marked
invalid). Fix would be in AddArpReq() when we process the arp rt, enqueue
processing for arp nh only when the existing entry is not found, else reuse
the old entry. This avoids arp nhs from being marked invalid.

Closes Jira Bug:JCB-218924

Change-Id: Ia6ffdfb3867278b8c7f71d99af8a012fc1e2cd81