Use <asn>:7999999 as the ethernet segment (es) route-target
Instead of using <asn>:1, use a number which is likely not already used in
the field. Define two numbers, one for 4 byte as and one for 2 byte.

For 2 byte global asn, <asn>:7999999 is used as the es route-target.
For 4 byte asn, <asn>:7999 is used. These values are 1 less than the
auto-rtarget value range beginning (8000000 and 8000)

Change-Id: Ibbdc30364cd77239c262c25874fc48190c832b43
Partial-Jira-Bug: CEM-3409