[config] Allow authentication even if the domain is not synced
That patch allows VNC API authentication even if the the domain where
the token is scoped is not synced in Contrail. This is not a problem for
an admin account (like juju uses in its deployments) as admin requests
by-pass all RBAC and perms checks but it's still an issue for non-admin
account. If the domain of an non admin account does not exists, RBAC
could fail to establish API access rules and perms checks could have
some issue to verify the access rights. This will be address in another
patch.

Change-Id: I0c2d32afcc96f37ea63bd4c3711350495ecf9198
Closes-Jira-Bug: JCB-219313