Fix for service health check assert
Delete and Add/Update events for service-health-check
are happening in the same second, when guest vm gets
restarted. Since Add/Update is processed when
service-health-check db-entry is marked for delete,
assert is happening in db-infra.
Ideally db-infra should have removed the delete mark for the
db-entry when Add/Update is processed. It has to be
root-caused. For now delete flag is removed from health-check
OperDBOnChange() and doing in health-check wont give any
side-effects.

Change-Id: I8a1540b9151459e3da2c2e70f22de80fdcb86299
Closes-jira-bug: CEM-3336
Author: Sangarshan Pillareddy <sangarshp@juniper.net>