[config] Clean VNC pre authetication code
Store the default domain and project dict as a property of the API server
instance to easily access it without doing a DB read.
Ensure authentication headers always have values.
Replace the default Keystone domain ID by the UUID of the default Contrail
domain.
Permits to define in configuration file, the value of the default Keystone
domain ID.

Change-Id: I82dd54b0fa8972f5e9f21798eb6bee1d94d9ecfd
Partial-Jira-Bug: CEM-3456
Closes-Jira-Bug: CEM-3691