[Config] Ignore ConcurrentObjectUseError in vnc_kombu
When either the publisher or drainer greenlet looses
connection to Kombu, it will call reconnect function to
establish a new connection.
Due to a timing issue, sometimes we end up in a situation
where one of those greenlets calls kombu APIs which results
in access of the same socket.
This results in ConcurrentObjectUseError by gevent.

The greenlet which got this exception will again try to
reconnect and this leads to a problem where both the
greenlets are not getting the socket long enough to process
any messages.

Using this change, we are now ignoring this Exception as
ConcurrentObjectUseError means there is already a valid
connection and we should not be doing a reconnect again.

Change-Id: I31abe33c2718debac27bf06e0905c65b3c1ee95e
Closes-Jira-Bug: JCB-219314