Enabled TCP Keep alive for Discovery Server
Currently, we don't enable TCP keepalive options for any of the
ports created by Gevent servers. This change is to enable TCP
Keepalive options for Discovery Server on port 9110.


For the other port 5997 that dicovery server opens, TCP KA options
are enabled as part of the commit https://review.opencontrail.org/#/c/50652/

Change-Id: If970e087264673e756dd61501bcaa7c58a8871ba
Closes-Jira-Bug: CEM-3186