[config] Fix tag ref validation
In patch Ie60fa525089aabb7e97e001fbbb48b4267ab0c39 we restrict the resource
tag references to be unique for most of the tag types. But Firewall Group
should be able to be to leverage that limitation as it can have the same
tag type in its endpoints. Code already prevents user to define tag ref to
firewall rule. The fix moves the tag ref validation after pre create/update
hooks as firewall rule ref are modified in that hook and it allows firewall
rule to have multiple tag refs to a same tag type.

Change-Id: Ibeda9e810601c6f64570fe1d5d7862ffa12a0b9d
Closes-Jira-Bug: CEM-4950