Read latest SG object after creating rules
When creating security group object, it first create a SG object and then
create 2 rules associated with that SG. However the associated rules was
not shown in CLI, because the create function return the original local SG
object dict, which has empty rule list and is not updated after creating
associated rules. This fix will read the latest SG object after creating
rules.

Change-Id: I7077d5368e3294229b55b60c0bd01113aad4d35c
Closes-Jira-Bug: CEM-4995