[config] Fix Keystone auth middleware
Since patch I82dd54b0fa8972f5e9f21798eb6bee1d94d9ecfd, a new config flag
was passed to the Keytone auth middleware as we though it will be ignored. But
in a specific case it breaks the Keyetone auth middleware. This patch removes
it to Keystone auth middleware and passed it to the necessary code only.
It also fix another typo introduce by the same patch.

Change-Id: I096fd5f1e1672be003e4bf5de8463c53c7637f4e
Closes-Jira-Bug: CEM-5098