Add a way to enable 4 byte as capability
Not all deployments need 4 byte AS support as typically contrail works within
data-center and uses a private asn (default: 64512). Hence disable 4 byte AS
feature by default. Add a knob to contrail-control.conf to enable it instead.

This also reduces entropy and risks to deployments that do not use 4 byte AS in
their networks.

Closes-Jira-Bug: CEM-5385

Conflicts:
	src/bgp/bgp_peer.cc
	src/bgp/bgp_server.cc
	src/bgp/bgp_server.h

Change-Id: Ide37729cdb86c65484706c822babefbcda720e34