Add a way to enable 4 byte as capability
Not all deployments need 4 byte AS support as typically contrail works within
data-center and uses a private asn (default: 64512). Hence disable 4 byte AS
feature by default. Add a knob to contrail-control.conf to enable it instead.

This also reduces entropy and risks to deployments that do not use 4 byte AS in
their networks.

Change-Id: I61d17bd862b7c45e0800471b5d8362f4be8ea7fb
Closes-Jira-Bug: CEM-5385