Update bgpaas rtarget routes when as number changes
For bgpaas, rtarget routes are auto-generated when ever bgpaas session comes
up. This rtarget route is in asn:target format. If server asn is changed in
the configuration, then all previously created rtarget routes must be deleted
and recreated using new as number.

Since this was not done, routes remained in the tables, causing peer path
cound assertion to fail later, during peer deletion.

TODO: Add unit test to cover this scenario. There is a system test which
      already found this particular issue though.

Change-Id: I44030e9e336d2b7097d0d940816e9fa72a1919dd
Partial-Jira-Bug: CEM-5450