Agent test framework changes
These changes allow the agent to run stand alone without having
a full contrail setup. The conf file parameters decide that the
agent is running in the mock mode. This feature is called "CAT"
which stands for "Contrail Agent Test" framework. This Agent can
run along with the Control node as well as the DPDK.

partial-jira-bug: CEM-5300
partial-jira-bug: CEM-6134

Change-Id: I2433740da521e1a9b095ec2db27fe9ead2c58cd0