[Config] Correct the references of draft firewall policy
We noticed that on deletion of a firewall policy, the draft firewall
policy had references to committed firewall rules.

Due to this, when a draft firewall policy is discarded, the node.js
sent deletion for these firewall resources.
1. Delete of draft firewall policy
2. Delete of committed firewall rule (which fails as still has back refs to committed firewall policy)

With this commit, we make sure the draft firewall policy marked for
deletion will have reference to only the draft firewall rules

Change-Id: Ibd9eace7d81d07ea5de262581f2b106a11f01d3c
Closes-Jira-Bug: CEM-6010