[config] Fix perms2 default values
In case of global or share access perms property miss in the
resource, take the default value to prevent code failing. Default global
access value is 0 (no global read, write or link) and the default share
access is set to an empty list of domain/project (shared with nobody).

Change-Id: I5b3946b4f66e36474449e9b625e5c998191c6e90
Close-Jira-Bug: CEM-6168