Modified to support diff cluster ids in a fabric
In schema transformer; bgp router peering function, cluster id is used
to determine if the bgp router is a route reflector(RR) or not.

We added the support for HA RR in ST that uses cluster id to determine
if RR needs to be peered but within a fabric RRs for HA needs to have
different cluster id.
This change adds the support to handle diff cluster id within a fabric.

Change-Id: I2e32e7d8741a22d6d0ea3847d76d4825162bd2ce
Closes-Jira-Bug: CEM-6769