updating enable_4b flag in bgp_server
flag was not getting updated based on config. Rather, flag is not needed
there since we can get it from global_system_config parameters.
Also removing the workaround that was added in control node
for correcting route target in case asn is 4 byte, so that auto
generated route target were getting corrected.

Closes Jira Bug: CEM-6992
Closes Jira Bug: CEM-7012

Change-Id: Ic7f2e1c526f4d15092b1821022fc27d45f0573e9