[Config] Get proper length for padding of key/password for physical router
We hit the issue where Config API was using rjust(32) on a password
that was 44 characters. rjust will just return 44 characters in this case.
This however will cause an issue for AES and it requires a password to be
multiple of 16 characters.

Changed code to pad both key and password for a physical router to next
higher multiple of 16 characters.

Closes-Jira-Bug: CEM-7083
Change-Id: I3964f8e6f4a832099e0029d7d9606403d966928b
(cherry picked from commit f7947e5e631a0adc90fbe9f2722f7852484d9c74)