Shrink FreeList container in case of overflow
In high-load scenarios with a huge flow rate the FreeList container may grow
rapidly, using too much RAM.
We actually do not need to store a lot of dummy flow entries in this list, so
it is safe to just shrink it whenever it reaches some threshold, kMaxThreshold.

In order not to let FreeList grow too much, we do the following.
Whenever the table FreeList size has reached kMaxThreshold (100 * 1000),
we do not add released entry back to FreeList in Free() method.
Instead, we just free the memory allocated for that entry.
Also, removed unused variable grow_count_.

Closes-Jira-Bug: TFB-1468
Related-Bug: #1571163
Change-Id: Idee56534ecf6362646c96d9e8d439abe08094cef