[config] Fix default quota management
Contrail permits to define default quotas for all projects in its
configuration file. Some part of the code did not taking care of that
default quota value if it was no defined in the project. That leads to
a bug where not defined quotas were constantly initialized and remove by API
server each time a project was updated (update, ref-update or
collection-update).

Change-Id: Id940c9064d825142002ee2aa667950a4befb35e3
Closes-Bug: CEM-7232