[Config] Handle the 'L' at the end of 4 byte ASN correctly.
When a Route Target has a 'L' appened to ASN field, it signifies that it's
a 4 byte ASN.
While storing this Route Target, we need to remove the 'L' and store in DB.
This was not being done before. This patch takes care of it in VirtualNetwork
user defined RTs and Logical Router user defined RTs.

Closes-Jira-Bug: CEM-7831
Change-Id: Ibaa405fa092a8d1a1e1035097d654fe3bb91567a