[Fabric] Read autonomous system configuration from the device during during device import
Read autonomous system configuration from the device during during import

This is needed for brownfield case. If device is configured with autonomous system,
Contrail does not push AS configuration on the device. If it is not configured,
contrail will push AS config.

Changing global-asn to brownfield-global-asn terminology.

Adding DM changes for abstract config and PR Cache to include asn values.

target-release: R1909
closes-jira-bug: CEM-7930
Change-Id: I8ea46774065ba4761c19fea059d47fe6a7b64074