This commit addresses the following two issues –
1. Deletion of EVPN_ROUTING peer paths of INET routes and EVPN type 5 routes
depend on VxlanRoutingManager getting route notification for EVPN type 5 and
type 2 route. It is possible that these notifications are not received on VRF
deletion if VRF deletion notification arrives before route notification to
VxlanRoutingManager. This is because, in VRF delete notification processing,
VxlanRoutingManager unregisters its route listeners. This can result in
EVPN_ROUTING peer paths and EVPN type 5 routes in routing vrf to linger
eventually resulting in VrfEntryDelete timeout crash. To address this, in
explicit delete route walker that gets called on route table delete, we clean up
lingering EVPN_ROUTING peer paths of routes and any EVPN type 5 route from
routing VRF.
2. Deletion of default route from INET table of bridge VRF was not happening
on VrfNotify in VxlanRoutingManager as the vrf’s vn would be reset to NULL by
that time. To address this, a flag is introduced in VxlanRoutingState to track
if the VRF is a bridge VRF or not. In HandleDefaultRoute, default route is
cleaned up if the VRF is a bridge VRF.

Closes-Jira-Bug: CEM-5861
Change-Id: I4578da827626b92aa60f7cbc586d5fda3666b2f8