[Config] Have different connections for publisher/drainer in API server
Some times, when both drainer and publisher greenlet try to access the same
socket, we see "ConcurrentObjectUseError" exception.
However, in a scaled setup, this is occuring very frequently.
This review is to explore the option of creating a seperate connection for
both publisher and drainer greentlet in API server

jira-bug: CEM-9273

Change-Id: I353676bee3885243de4086a0fdf0c991e189d8c0