[Config] Have different connections for publisher/drainer in API server
Some times, when both drainer and publisher greenlet try to access the same
socket, we see "ConcurrentObjectUseError" exception.
However, in a scaled setup, this is occuring very frequently.
This review is to explore the option of creating a seperate connection for
both publisher and drainer greentlet in API server

Jira-bug: CEM-9273
Change-Id: I89ee275fd39058c3cbe0c3255906d5164dbc2d00