[Config] Have different connections for publisher/drainer in API server
Some times, when both drainer and publisher greenlet try to access the same
socket, we see "ConcurrentObjectUseError" exception.
However, in a scaled setup, this is occuring very frequently.
This review is to explore the option of creating a seperate connection for
both publisher and drainer greentlet in API server

jira-bug: CEM-9273
Change-Id: I302475de715dda12a4b1c6786e94f25d9944cd7a