[DM] Security group defined on virtual port group is applied to all interfaces of same virtual network
The firewall filter is generated not only for that particular interface/VLAN but
also for all interfaces/VLANs which belong to the same Virtual Network in that router.
In other words, is its being applied to other VPGs assigned to the same virtual network
in the same router, even though these other VPGs don't have the Security Group applied.

Modifying the VPGs assignment based on vlan tag and interface. Also handled
old way of role invocation.

Change-Id: I4faa5d6ca7077082319a807e6ec903de6606674a
target-release: R1909
closes-jira-bug: CEM-8199