[DM] Upon Power Off/Power On of CEM - Configuration is pushed on EVPN/VXLAN Fabric Devices
The autonomous system number and peer asn are reassigned to a new value after docker restart.
We store the autonomous system in the Cassandra db and read it back for consistency. However,
since asn is a number, the value was not getting stored in the cassandra db as integer values
cannot be encoded. Therefore, we now stringify the asn before storing it in Cassandra db and
convert it into an integer after db read. This ensures that for devices, that have an asn already,
no new asn is being reassigned on docker restart.

Change-Id: Ic0c759c93cdd01d0e1310b7292ef42db27de46a8
target-release: R1909
closes-jira-bug: CEM-7511