[Config] Added 4 byte ASN check for local-as of bgpaas object
The 4byte ASN checks were not done at API server for local-as field
of the bgpaas object. This review takes care of that.
We also add restrictions to disallow update of both 'enable_4byte_as' flag
and 'ASN' in same request for global_system_config.
Removed that restrictions as well

Closes-Jira-Bug: CEM-7880
Change-Id: I596c0945668e00ca7ece0c10be58bf4a88f06ec2