Identify the service-instance(si) from fq_name instead of vmi name
vmi name for logical-router and lbaas are created by joining
domain-name, project-name, service-instance(si) name and etc
with '__'.  During vmi update, vmi name is used to find the si
inorder to create vm(netns instance) and link to si if the vm
and si link is not done properly. if domain-name or project-name
has any '__', the exception would happen due to the current logic
to get the interface index to form the vm name. Instead of finding
si from vmi name, si fq_name is directly used.

closes-jira-bug: CEM-7788
Change-Id: Ifca3a79402f0e3f65cd603bc078715e2879aaabd