The fix implements two changes for flows matching a VN with "IP Fabric Forwarding" enabled:
(1) Before the fix, if the source IP address matches a subnet IPAM,
then the flow is only considered valid if it there is an overlay
route to the source. This condition is now relaxed in the case
where the subnet IPAM matches the vhost0 subnet.

(2) Before the fix, a flow with the characteristics specified in (1)
is forwarded in the context of the VN it belongs to, rather than
in the fabric VRF, even though the VN has "IP Fabric Forwarding"
enabled. This is changed now so such a flow is forwarded via inet.0.

target-release: R1909
closes-jira-bug: CEM-7840
Change-Id: Ia4e7767e141dc648bfc802004a0098d232d83e7e