[Config] Added code to skip vpg verifications when VMI updated without any vlan changes
Most of the VPG verifications done CEM-6767 are done for VMI create only.
However, the code is also invoked as part of update as part of a common code.
Because of these verifications, some of the VMI updates are failing.
We should skip VPG verifications in cases when there was no update to VLAN for a VMI.

closes-jira-bug: CEM-8508
Change-Id: I0a34dfed37b919c24efc42bed164e6c3b1d50c7b