[VNC OpenStack] Fix firewall rule IPv6 support
Don't use any more the FR 'any' endpoint property when a rule address is
a wildcard (0/0 or ::/0) but set it in the 'subnet' endpoint property.
Deduce the FR IP version from the 'subnet' endpoint properties and use
version 4 by default.
Limit FR to have only one FG as source or destination which permits to
deduce the IP version of the rule.
If a FR protocol is set to ICMP, convert it to IPv6-ICMP if needed and same
for the reverse.

Change-Id: I9c93821cd326edc422936119332748ecedadb364
Closes-Jira-Bug: CEM-8738