Vrouter agent is initializing
There is a OnChange function in physical Interface which was never
getting hit, so after bond monitoring changes that function
gets hit. In that function it is setting vrf name as NULL
due to which any change in bond interface state it sets vrf as
NULL and it leads to invalid next hop.
Thus agent gets into initialize state.
Added a null check.
Also corrected the code alignment of some files.
Added test case for this use case.

closes-jira-bug: CEM-9024
Change-Id: I0c3082206468c486081274f6e12a4a41da6dcf3b