Modified to support diff cluster ids in a fabric
In schema transformer; bgp router peering function, cluster id is used
to determine if the bgp router is a route reflector(RR) or not.

We added the support for HA RR in ST that uses cluster id to determine
if RR needs to be peered but within a fabric RRs for HA needs to have
different cluster id.
This change adds the support to handle diff cluster id within a fabric.

Closes-Jira-Bug: CEM-6769
Closes-Jira-Bug: CEM-7749
Change-Id: I6104b59a9ed9bd3f9e611406b40b1ff6dfd6d453