Check to ignore deleted VM interface stats response from vrouter
Agent has crashed when it was processing the vm interface stats ksync reply
from vrouter at the same time when the vm interface was deleted. Uve module
receives vm interface delete callback and it marks the corresponding uve entry
as deleted and removes the vm reference from the entry.
A timer callback then removes the uve entry from the tree and subsequently
the entry is freed.
While processing the vrouter stats response for vm interface, additional
check is added to make sure we discard the response if it is already deleted.

Closes Jira Bug: CEM-9074

Change-Id: I4a0b8001c2c4460dc838742d6ec12c838c1c076b