[Fabric] Adding retry logic for fully loaded hardware
Fully loaded hardware like 10k8s and 10k16s have humoungous
day 0 configuration. When these zeroized devices come up, they
take a long time to load the config which results in exclusive lock
being set for the config mode.

We are fixing this problem by this:

1. creating a new role to attempt to open the config mode. If the
attempt is not successful, we are retrying for 5 mins.
2. Also for RMA cases, we are including the same retry so that for
such big configs, it does not attempt to push the supplemental config
while the override is still on-going.

Change-Id: I4c554e00b9964083ae4dcdb8478a675f28b97457
target-release: R1911
closes-jira-bug: CEM-8263