[Fabric] Adding retry logic for fully loaded hardware
Fully loaded hardware like 10k8s and 10k16s have humoungous
day 0 configuration. When these zeroized devices come up, they
take a long time to load the config which results in exclusive lock
being set for the config mode.

We are fixing this problem by this:

1. creating a new role to attempt to open the config mode. If the
attempt is not successful, we are retrying for JUNOS_COMMIT_TIMEOUT set
in all.yml.
2. Also for RMA cases, we are including the same retry so that for
such big configs, it does not attempt to push the supplemental config
while the override is still on-going.

Both these cases can be handled by a pilot config mode check
before beginning of the assign static ip workflow where config mode
is opened as part of read interfaces info. Also, once after the config
override is done, we may still have pending commits that may take time
to complete. Therefore including a check at the end of the assign static
ip workflow as well.

The JUNOS_COMMIT_TIMEOUT is meant as a configurable timeout value so that
it can be increased to bigger timeouts for 10k16 devices with even huge configs.

Change-Id: I4c554e00b9964083ae4dcdb8478a675f28b97457
target-release: R1911
closes-jira-bug: CEM-8263