Protect uve entry to prevent access after reset
This issue is seen when a VM is shutdown/start for around 250 times.
The uve module in agent receives delete notification for a VM interface
and marks the uve entry as deleted and resets the information.
Periodic ksync replies from vrouter are received and updated in the uve
entry.
A race condition between the above to threads results in accessing the uve
entry after it is marked for delete.
Re-arranging the uve entry mutex to protect against this race condition.
Tested this fix on the setup with 250 VM restarts and no crashes reported.
Since this is a corner case scenario, the existing UT cases which covers
this part of the code is unable to catch this scenario.

Closes Jira Bug: CEM-9074

Change-Id: Ic509b7982b1fc66b8104e5bbd306fe59bf421bcc