Add assertion to ensure that only IFMapNode is added to ifmap link table
There are some data which points to weird information that IFMapLink data
inserted as IFMapNode data in IFMAP link table. This should never happen,
but a few core files indicate so.

Add a simple assertion in code to catch this, (if at all it does happen)
asap when the entry is added into the table (as oppose to detecting much
later when most of the context of insertion is lost)

Closes-Jira-Bug: CEM-9662
Change-Id: Ia20ee478737e4447301dbc79786eedc8393745c6