Enabling healthchecks for native LBaaS driver leads to vrouter crash
Issue:
-----
The crash is happening because we are asserting when we get exception
out_of_range for substr operation from heathcheck.cc
In BuildData function the url_path is coming as "/", so this is
passed to http_parser_parse_url to form urldata. In this url_data
url field is set as UF_PATH. But using substr we are trying to get
data for url field UF_HOST which has some default value set in
urldata.field_data[UF_HOST].off which is obviously greater than
url_path length i.e 1 and hence this throws an exception.

Fix:
----
Check that the url field is set to UF_HOST before calling substr.

To fix this, we need to check offset length should not be greater than the
length of string.

closes-jira-bug: CEM-9813
Change-Id: I906b65608e57e6a53080261b5c8b0f0edde346c4